backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-html>=4.1.1",
]

[tool.pytest.ini_options]
testpaths = ["../tests/backend"]
pythonpath = ["."]
addopts = "--html=../tests/reports/backend_report.html --self-contained-html -q"
